+++ src/mjlab/asset_zoo/robots/p1/__init__.py
@@ -1,0 +1,8 @@
+"""P1 biped robot."""
+
+from mjlab.asset_zoo.robots.p1.p1_constants import (
+  P1_ACTION_SCALE as P1_ACTION_SCALE,
+)
+from mjlab.asset_zoo.robots.p1.p1_constants import (
+  get_p1_robot_cfg as get_p1_robot_cfg,
+)